introduce playpen package
