+++ clemcore/playpen/__init__.py
@@ -1,0 +1,13 @@
+from clemcore.playpen.callbacks import BaseCallback, GameRecordCallback, RolloutProgressCallback, CallbackList
+from clemcore.playpen.base import BasePlayPen
+from clemcore.playpen.env import GameEnv, make_env
+
+__all__ = [
+    "BaseCallback",
+    "GameRecordCallback",
+    "RolloutProgressCallback",
+    "CallbackList",
+    "BasePlayPen",
+    "GameEnv",
+    "make_env"
+]
